# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Notification Service API
  description: |
    # Notification Service API Documentation
    
    A comprehensive notification service that supports multiple delivery channels (Email, SSE, Push) 
    with template management and user preferences.
    
    ## Features
    - 📧 **Email Notifications** - Send templated emails via SMTP
    - 🔔 **Real-time SSE** - Server-Sent Events for instant notifications
    - 📱 **Push Notifications** - (Coming soon)
    - 🎨 **Template Management** - Dynamic templates with variable substitution
    - ⚙️ **User Preferences** - Customizable notification settings
    - 🔄 **Event-Driven** - Redis Streams for reliable event processing
    - 🔁 **Retry Mechanism** - Automatic retry for failed notifications
    
    ## Authentication
    All endpoints (except health check) require JWT Bearer token authentication from Casdoor OAuth2.
    
    ## Base URL
    - **Development**: `http://localhost:8082/notification-service`
    - **Production**: `https://api.yourcompany.com/notification-service`

  version: 2.0.0
  contact:
    name: Notification Service Team
    email: support@examplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8082/notification-service
    description: Development server
  - url: https://api.yourcompany.com/notification-service
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/notifications:
    get:
      tags:
        - User Notifications
      summary: Get notification history
      description: Get paginated notification history for the current user (PUSH notifications only)
      operationId: getNotificationHistory
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated notification history
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserNotification'
                  pageable:
                    $ref: '#/components/schemas/Pageable'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/notifications/unread-count:
    get:
      tags:
        - User Notifications
      summary: Get unread count
      description: Get the count of unread notifications for the current user
      operationId: getUnreadCount
      responses:
        '200':
          description: Unread notification count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 5
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/notifications/{id}/read:
    put:
      tags:
        - User Notifications
      summary: Mark notification as read
      description: Mark a specific notification as read
      operationId: markNotificationAsRead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/notifications/mark-all-read:
    put:
      tags:
        - User Notifications
      summary: Mark all notifications as read
      description: Mark all notifications as read for the current user
      operationId: markAllNotificationsAsRead
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  markedCount:
                    type: integer
                    example: 10
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/notifications/{id}:
    delete:
      tags:
        - User Notifications
      summary: Delete notification
      description: Delete a specific notification
      operationId: deleteNotification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Notification deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== USER PREFERENCES ==
...
  /api/v1/sse/connect:
    get:
      tags:
        - SSE
      summary: Connect to SSE stream
      description: |
        Establish a Server-Sent Events connection to receive real-time notifications.
        
        **Note**: You can pass JWT token either as:
        - Authorization header: `Bearer YOUR_TOKEN`
        - Query parameter: `?token=YOUR_TOKEN` (for browser EventSource compatibility)
        
        Keep this connection open to receive events.
      operationId: connectSSE
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
          description: JWT token (alternative to Authorization header)
      responses:
        '200':
          description: SSE connection established
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                connectionEvent:
                  value: |
                    event: connect
                    data: {"message":"Connected to notification service","userId":123,"timestamp":"2025-11-23T10:00:00Z"}

                notificationEvent:
                  value: |
                    event: notification
                    data: {"type":"assessment.published","content":"<html>...</html>","timestamp":"2025-11-23T10:01:00Z"}

                heartbeatEvent:
                  value: |
                    event: heartbeat
                    data: {"timestamp":"2025-11-23T10:02:00Z","type":"heartbeat"}
  /api/v1/sse/subscribe/{topic}:
    get:
      tags:
        - SSE
      summary: Subscribe to topic
      description: Subscribe to a specific topic for broadcast messages
      operationId: subscribeToTopic
      parameters:
        - name: topic
          in: path
          required: true
          schema:
            type: string
          example: assessment-updates
        - name: token
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Subscribed to topic
          content:
            text/event-stream:
              schema:
                type: string
  /api/v1/sse/status/{userId}:
    get:
      tags:
        - SSE
      summary: Check connection status
      description: Check if a user is connected to SSE
      operationId: checkUserSSEStatus
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Connection status
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  connected:
                    type: boolean
                  timestamp:
                    type: integer
...
  /api/v1/sse/test/send-to-user:
    post:
      tags:
        - SSE
      summary: Send test notification to user (Admin)
      description: Send a test notification to a specific user via SSE. Also saves to database as PUSH notification. (Admin only)
      operationId: sendTestNotificationToUser
      parameters:
        - name: targetUserId
          in: query
          required: true
          schema:
            type: integer
        - name: message
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            default: test_notification
        - name: subject
          in: query
          required: false
          schema:
            type: string
            default: Test Notification
      responses:
        '200':
          description: Test notification sent and saved to database
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  notificationId:
                    type: string
                    format: uuid
                  savedToDb:
                    type: boolean
                  sentViaSse:
                    type: boolean
                  targetUserId:
                    type: integer
                  message:
                    type: string
                  timestamp:
                    type: integer
  /api/v1/sse/test/broadcast:
    post:
      tags:
        - SSE
      summary: Broadcast to topic (Admin)
      description: Broadcast a message to all users subscribed to a topic (Admin only)
      operationId: broadcastToTopic
      parameters:
        - name: topic
          in: query
          required: true
          schema:
            type: string
        - name: message
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message broadcasted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  topic:
                    type: string
                  message:
                    type: string

  # ==================== EVENT PUBLISHING ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from Casdoor OAuth2.
        
        To get a token, authenticate with Casdoor:
        ```
        POST https://casdoor.sapx.me/api/login/oauth/access_token
        Content-Type: application/x-www-form-urlencoded
        
        grant_type=password&
        username=YOUR_USERNAME&
        password=YOUR_PASSWORD&
        client_id=YOUR_CLIENT_ID&
        client_secret=YOUR_CLIENT_SECRET
        ```
  schemas:
...
    UserNotification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          example: assessment.published
        subject:
          type: string
          example: New Assessment Available
        content:
          type: string
          example: You have been assigned a new assessment...
        isRead:
          type: boolean
          default: false
        createdAt:
          type: number
          description: Timestamp in epoch seconds
          example: 1765300386.744872

    # ==================== PREFERENCES ====================
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
        totalElements:
          type: integer
        sort:
          type: object

    # ==================== ERRORS ====================
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            timestamp: "2025-11-23T10:00:00Z"
            status: 401
            error: Unauthorized
            message: Full authentication is required to access this resource
            path: /api/v1/templates
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            timestamp: "2025-11-23T10:00:00Z"
            status: 404
            error: Not Found
            message: Template not found
            path: /api/v1/templates/unknown_template